exercise-1_6/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0 